Moved slide and other utility driving methods out of MecanumDrive for ease of updating :)
